Deal with numerical boundaries for particle probablity(weights)

+++ src/helper_functions.h
@@ -59,9 +59,8 @@ inline double dist(double x1, double y1, double x2, double y2) {
   return sqrt((x2 - x1) * (x2 - x1) + (y2 - y1) * (y2 - y1));
 }
 
-
 // Multivariate Gaussian
-inline long double multiv_prob(double sig_x, double sig_y, double x_obs, double y_obs,
+inline double multiv_prob(double sig_x, double sig_y, double x_obs, double y_obs,
                    double mu_x, double mu_y) {
   // calculate normalization term
   double gauss_norm;
@@ -73,7 +72,7 @@ inline long double multiv_prob(double sig_x, double sig_y, double x_obs, double 
                + (pow(y_obs - mu_y, 2) / (2 * pow(sig_y, 2)));
     
   // calculate weight using normalization terms and exponent
-  long double weight;
+  double weight;
   weight = gauss_norm * exp(-exponent);
     
   return weight;
